+++ src/app/core/store/core-store.module.ts
@@ -4,15 +4,21 @@ import { environment } from '../../../environments/environment';
 
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
+import { RouterEffects } from './router/router.effects';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+import { StoreRouterConnectingModule } from '@ngrx/router-store';
+import { routerReducers, RouterStateSerializerProvider } from './router/router.state';
 
 @NgModule({
-  declarations: [],
   imports: [
     CommonModule,
-    StoreModule.forRoot({}),
-    EffectsModule.forRoot([]),
+    StoreModule.forRoot(routerReducers),
+    EffectsModule.forRoot([RouterEffects]),
+    StoreRouterConnectingModule.forRoot(),
     !environment.production ? StoreDevtoolsModule.instrument() : [],
+  ],
+  providers: [
+    RouterStateSerializerProvider
   ]
 })
 export class CoreStoreModule { }