+++ src/app/modules/admin/guards/admin-activation.guard.ts
@@ -1,7 +1,11 @@
 import { Injectable } from '@angular/core';
-import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
+import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 
 import { Observable } from 'rxjs';
+
+import { Store } from '@ngrx/store';
+import { AppState } from '../../../core/store/app.state';
+import * as RouterActions from '../../../core/store/router/router.actions';
 
 import { AppSettingsService } from '../../../core/services/app-settings.service';
 
@@ -10,7 +14,7 @@ import { AppSettingsService } from '../../../core/services/app-settings.service'
 })
 export class AdminActivationGuard implements CanActivate {
   constructor(
-    private router: Router,
+    private store: Store<AppState>,
     private appSettingsService: AppSettingsService
   ) {}
   canActivate(
@@ -19,7 +23,9 @@ export class AdminActivationGuard implements CanActivate {
     if (this.appSettingsService.getAppSettings().mode === 'admin') {
       return true;
     } else {
-      this.router.navigate(['/']);
+      this.store.dispatch(new RouterActions.Go({
+        path: ['/']
+      }));
     }
     return false;
   }