+++ src/app/modules/cart/components/cart-modal/cart-modal.component.ts
@@ -1,15 +1,17 @@
 import {
   OnInit,
+  DoCheck,
   Component,
-  ChangeDetectionStrategy, DoCheck, OnDestroy,
+  OnDestroy,
+  ChangeDetectionStrategy
 } from '@angular/core';
-import { Router } from '@angular/router';
 
 import { Observable, Subscription } from 'rxjs';
 
 import { select, Store } from '@ngrx/store';
 import { AppState } from '../../../../core/store/app.state';
-import { CartState } from '../../../../core/store/cart/cart.state';
+import * as CartSelectors from '../../../../core/store/cart/cart.selectors';
+
 import {
   RemoveProductFromCart,
   DecreaseProductQuantity,
@@ -17,6 +19,7 @@ import {
 } from '../../../../core/store/cart/cart.actions';
 
 import { Product } from '../../../products/models/product.model';
+import * as RouterActions from '../../../../core/store/router/router.actions';
 
 @Component({
   selector: 'app-cart-modal',
@@ -26,7 +29,10 @@ import { Product } from '../../../products/models/product.model';
 })
 export class CartModalComponent implements OnInit, DoCheck, OnDestroy {
 
-  cartState$: Observable<CartState>;
+  cartCost$: Observable<number>;
+  cartEmpty$: Observable<boolean>;
+  cartProducts$: Observable<Product[]>;
+
   cartProducts: Product[];
   lastUpdated: number;
   sortingOrder: boolean;
@@ -35,15 +41,16 @@ export class CartModalComponent implements OnInit, DoCheck, OnDestroy {
   private sub: Subscription;
 
   constructor(
-    private store: Store<AppState>,
-    private router: Router
+    private store: Store<AppState>
   ) { }
 
   ngOnInit() {
-    this.cartState$ = this.store.pipe(select('cart'));
+    this.cartCost$ = this.store.pipe(select(CartSelectors.getCartCost));
+    this.cartEmpty$ = this.store.pipe(select(CartSelectors.getCartEmpty));
+    this.cartProducts$ = this.store.pipe(select(CartSelectors.getCartProducts));
 
-    this.sub = this.cartState$.subscribe(
-      cartState => this.cartProducts = cartState.products
+    this.sub = this.cartProducts$.subscribe(
+      productsList => this.cartProducts = productsList
     );
   }
 
@@ -72,7 +79,9 @@ export class CartModalComponent implements OnInit, DoCheck, OnDestroy {
   }
 
   order(): void {
-    this.router.navigate(['/order']);
+    this.store.dispatch(new RouterActions.Go({
+      path: ['/order']
+    }));
   }
 
   getSortingOrder(): string {