+++ src/app/modules/products/components/product-feedback/product-feedback.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 
-import { Router } from '@angular/router';
+import { Store } from '@ngrx/store';
+import { AppState } from '../../../../core/store/app.state';
+import * as RouterActions from '../../../../core/store/router/router.actions';
 
 @Component({
   selector: 'app-product-feedback',
@@ -10,13 +12,15 @@ import { Router } from '@angular/router';
 export class ProductFeedbackComponent implements OnInit {
 
   constructor(
-    private router: Router
+    private store: Store<AppState>
   ) { }
 
   ngOnInit() {
   }
 
   close(): void {
-    this.router.navigate([{outlets: { feedback: null }}]);
+    this.store.dispatch(new RouterActions.Go({
+      path: [{outlets: { feedback: null }}]
+    }));
   }
 }